+++ Java/Eclipse-Workspace/CalculadoraDAM/src/tests/CalculadoraAbsTest.java
@@ -11,16 +11,16 @@ public class CalculadoraAbsTest extends CalculadoraBaseTest {
 	@Test
 	public void absAcierto() {
 		double a = 5;
-		int esperado = 5;
-		int resultado = Calculadora.opAbs(a);
+		double esperado = 5;
+		double resultado = Calculadora.opAbs(a);
 		assertEquals(esperado, resultado);
 	}
 	
 	@Test
 	public void absFallo() {
-		int a = 5;
-		int esperado = 15;
-		int resultado = Calculadora.opAbs(a);
+		double a = 5;
+		double esperado = 15;
+		double resultado = Calculadora.opAbs(a);
 		assertEquals(esperado, resultado);
 	}
